predict_transmitter: fix issue with string roots

+++ fafbseg/flywire/synapses.py
@@ -217,7 +217,7 @@ def predict_transmitter(x, single_pred=False, weighted=True, mat='auto',
     pred = collapse_nt_predictions(syn, single_pred=single_pred,
                                    weighted=weighted, id_col='pre')
 
-    return pred.reindex(make_iterable(x), axis=1)
+    return pred.reindex(make_iterable(x).astype(int), axis=1)
 
 
 def fetch_synapses(x, pre=True, post=True, attach=True, min_score=30, clean=True,
